moved hash and address
