+++ lib/contracts/contributionreward.ts
@@ -1,14 +1,13 @@
 "use strict";
 import dopts = require("default-options");
+import { Address, Hash } from "../commonTypes";
 
 import {
-  Address,
   ArcTransactionDataResult,
   ArcTransactionProposalResult,
   ArcTransactionResult,
   EventFetcherFactory,
   ExtendTruffleContract,
-  Hash,
   StandardSchemeParams,
 } from "../ExtendTruffleContract";
 import { Utils } from "../utils";
