+++ lib/contracts/genesisProtocol.ts
@@ -1,13 +1,12 @@
 "use strict";
 import dopts = require("default-options");
+import { Address, Hash } from "../commonTypes";
 import {
-  Address,
   ArcTransactionDataResult,
   ArcTransactionProposalResult,
   ArcTransactionResult,
   EventFetcherFactory,
   ExtendTruffleContract,
-  Hash,
 } from "../ExtendTruffleContract";
 import { Utils } from "../utils";
 const SolidityContract = Utils.requireContract("GenesisProtocol");
