+++ lib/contracts/upgradescheme.ts
@@ -1,13 +1,11 @@
 "use strict";
 import dopts = require("default-options");
-
+import { Address, Hash } from "../commonTypes";
 import {
-  Address,
   ArcTransactionDataResult,
   ArcTransactionProposalResult,
   EventFetcherFactory,
   ExtendTruffleContract,
-  Hash,
   StandardSchemeParams,
 } from "../ExtendTruffleContract";
 import { Utils } from "../utils";
